Fix generic vector template issue
